requirements.txt
torch
torchvision
super-image
pillow
streamlit
opencv-python
